python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openmc-metal"
version = "0.1.0"
description = "Monte Carlo neutron transport on Apple Metal GPU (Python)"
requires-python = ">=3.9"
dependencies = [
    "pyobjc-framework-Metal>=10.0",
    "pyobjc-framework-MetalKit>=10.0",
    "numpy>=1.24",
    "matplotlib>=3.7",
    "reportlab>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-timeout>=2.0",
]

[project.scripts]
openmc-metal-bench = "openmc_metal.benchmarks.runner:main"
openmc-metal-report = "openmc_metal.report.pdf_report:main"

[tool.setuptools.packages.find]
where = ["src"]
